refactor(globe-polygon): ラベル名の接頭辞重複解消／resolveTerrainEngine 共通化
- globePolygonManager: ラベルの DynamicTexture / Plane / Material 名が
  `globe-polygon-${id}-...`（id 自体が globe-polygon-N のため二重接頭辞）に
  なっていたのを `${id}-...` に統一し、他メッシュ名（${id}-point-... 等）と揃える。
- resolveTerrainEngine を terrain/urlState.ts に集約。viewer デモは再 export で
  既存 import 互換を維持し、polygon デモは urlState から import（実装重複を解消）。
